Moved initial fetch to be saved in the store
